Split SPM functionality into a separate collection; makes it easier to use outside STK
Also adding in these major changes:
* Group both importer and exporter as one module
* Bump version numbers for both importer and exporter to 2.0
* Add author information
