+++ .gitignore
@@ -1,0 +1,8 @@
+/__pycache__
+/.vscode
+/tmp
+
+**/*.pyc
+
+tmp.py
+verify.py
